Fix compiler warnings

+++ arch/x86/monitor/rawfs.c
@@ -136,7 +136,7 @@ void r_stat(ino_t inum, struct stat *stp)
     d2_inode *dip;
     int i;
 
-    dip= &blockbuf->b__v2_ino[ino_offset];
+    dip= &blockbuf->b__v2_ino[(unsigned int)ino_offset];
 
     curfil.i_mode= dip->d2_mode;
     curfil.i_nlinks= dip->d2_nlinks;
@@ -152,7 +152,7 @@ void r_stat(ino_t inum, struct stat *stp)
     d1_inode *dip;
     int i;
 
-    dip= &blockbuf->b__v1_ino[ino_offset];
+    dip= &blockbuf->b__v1_ino[(unsigned int)ino_offset];
 
     curfil.i_mode= dip->d1_mode;
     curfil.i_nlinks= dip->d1_nlinks;
